webapp/requirements.txt
# Flask web framework
Flask==3.0.0
Werkzeug==3.0.1

# Core dependencies from main project
langextract>=0.1.0
pyyaml>=6.0
numpy>=1.20.0

# Optional for PDF generation
reportlab>=3.6.0

# Production server
gunicorn==21.2.0

# Environment management
python-dotenv==1.0.0